refactor(tui-node): apply safe clippy --fix pass (24 fixes)
Batched automatic fixes via:
    cargo clippy --fix --lib -p tui-node --tests --allow-dirty -- \\
        -A clippy::all \\
        -W clippy::useless_vec \\
        -W clippy::useless_format \\
        -W clippy::unnecessary_cast \\
        -W clippy::needless_borrow \\
        -W clippy::manual_map \\
        -W clippy::unnecessary_lazy_evaluations \\
        -W clippy::collapsible_else_if \\
        -W clippy::collapsible_if

Lint-by-lint count from \`cargo clippy\` before → after:

- useless_vec:              7  → 0   (\`vec![...]\` → \`[...]\` for
                                       static-known arrays that only
                                       need .join())
- useless_format:           3  → 0   (\`format!("literal")\` →
                                       \`"literal".to_string()\`)
- collapsible_if:          34  → 19  (nested \`if let A && if let B\`
                                       merged via let-chains)
- map_or (manual_map /      ...
  unnecessary_lazy_eval):  10  → 3

9 files touched, 45 lines net deleted. All changes are semantically
equivalent and the surface they touch has unit tests; \`cargo test
--workspace --lib --tests --exclude mpc-wallet-native\` = 174
passing, unchanged.

Deliberately excluded this pass:
- clippy::all (blanket enable would include style lints that
  have produced bad \`--fix\` output before — derivable_impls
  reverted last tick, plus the MutexGuard-across-await lint
  which is non-mechanical and needs judgment).
- clippy::needless_borrowed_reference and similar — not in the
  20 \`the borrowed expression implements the required traits\`
  warnings; those remain as a candidate for a later pass.
- clippy::or_fun_call / or_insert_with (11 remaining) — these
  can eagerly evaluate closures in ways that change behavior
  (e.g. allocating when the entry is already present). Worth
  a dedicated review rather than a blanket fix.

+++ apps/tui-node/src/keystore/extension_compat.rs
@@ -244,7 +244,7 @@ impl ExtensionKeyShareData {
             session_id: wallet_data.session_id.clone(),
             device_id: wallet_data.device_id.clone(),
             participant_index: (participant_index + 1) as u16, // 1-based in extension
-            threshold: wallet_info.threshold as u16,
+            threshold: wallet_info.threshold,
             total_participants: wallet_info.total_participants,
             participants: wallet_info.devices.iter().map(|d| d.device_id.clone()).collect(),
             curve: curve.to_string(),
